Moved failure watcher to services package.
